+++ packages/bluesky/lib/bluesky.dart
@@ -129,7 +129,6 @@ export 'package:bluesky/src/entities/reason.dart';
 export 'package:bluesky/src/entities/reply.dart';
 export 'package:bluesky/src/entities/reply_ref.dart';
 export 'package:bluesky/src/entities/reposted_by.dart';
-export 'package:bluesky/src/feeds/feed_algorithm.dart';
 export 'package:bluesky/src/feeds/feeds_service.dart';
 export 'package:bluesky/src/graphs/graphs_service.dart';
 export 'package:bluesky/src/notifications/notifications_service.dart';